api/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
httpx>=0.25.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.28.0

# AI/ML
openai>=2.0.0  # Updated for GPT-5 support (responses.create API)
nltk>=3.8.1
pandas>=2.1.1
numpy>=1.26.0

# Caching
redis>=5.0.1
hiredis>=2.2.3

# Security
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
httpx>=0.25.0

# Monitoring and Logging
loguru>=0.7.0
prometheus-fastapi-instrumentator>=6.1.0

# Development
black>=23.9.1
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.1